Modularise server route files for independent pages
