+++ src/data/favourites.json
@@ -5,5 +5,12 @@
     "overview": "The wildly funny and touching story of a lonely Hawaiian girl and the fugitive alien who helps to mend her broken family.",
     "release_date": "2025-05-17",
     "backdrop_path": "/7Zx3wDG5bBtcfk8lcnCWDOLM4Y4.jpg"
+  },
+  {
+    "id": 372058,
+    "title": "Your Name.",
+    "overview": "High schoolers Mitsuha and Taki are complete strangers living separate lives. But one night, they suddenly switch places. Mitsuha wakes up in Taki’s body, and he in hers. This bizarre occurrence continues to happen randomly, and the two must adjust their lives around each other.",
+    "release_date": "2016-08-26",
+    "backdrop_path": "/8x9iKH8kWA0zdkgNdpAew7OstYe.jpg"
   }
 ]
